+++ copy_with_extension_gen/lib/src/copy_with_generator.dart
@@ -17,7 +17,7 @@ import 'package:source_gen/source_gen.dart'
 class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
   CopyWithGenerator(this.settings) : super();
 
-  Settings settings;
+  final Settings settings;
 
   /// Generates the `copyWith` extension code for the annotated [element].
   ///
@@ -82,6 +82,7 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
       skipFields: classAnnotation.skipFields,
       copyWithNull: generateCopyWithNull,
       constructor: resolvedConstructorName,
+      settings: settings,
       superInfo: superInfo,
     );
   }
@@ -100,7 +101,7 @@ class CopyWithGenerator extends GeneratorForAnnotation<CopyWith> {
     ClassElement element,
     CopyWithAnnotation annotation,
   ) {
-    var superInfo = findAnnotatedSuper(element);
+    var superInfo = findAnnotatedSuper(element, settings);
     if (annotation.skipFields &&
         superInfo != null &&
         element.supertype?.element != superInfo.element) {